--- client/src/components/ProjectsPage/Listing.scss
@@ -1,53 +1,0 @@
-@import "../Application.scss";
-
-//Project card
-.project-card{
-    display: flex;
-    overflow: 1 1 auto;
-    border-radius: 20px;
-    border: none;
-    box-shadow: 8px 13px 23px -5px rgba(0,0,0,0.5);
-
-    .project-card-image{
-        display: block;
-        max-width: 30%;
-        max-height: 90%;
-        float: left;
-        margin: auto 10px;
-        img{
-            background-color: white;
-            border-radius: 20px;
-        }
-    }
-
-    .project-card-body{
-        padding: 20px;
-        width: 100%;
-    }
-
-    .view-button{
-        z-index: 3;
-        position: relative;
-        border-radius: 20px;
-        border: none;
-    }
-
-    .githubIcon{
-        border-radius: 20px;
-        margin-right: 10px;
-        width: 140px;
-        z-index: 1;
-        position: relative;
-        left: -105px;
-
-        svg{
-            float: right;
-        }
-    }
-}
-.project-card.bg-0{
-    background-color: #B9E8B7;
-}
-.project-card.bg-1{
-    background-color: #A1E6EF;
-}
