+++ client/src/components/layout/Listing.js
@@ -4,7 +4,7 @@ import React from 'react';
 import {Card, Button} from 'react-bootstrap'
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import {faGithub} from '@fortawesome/free-brands-svg-icons'
-import './Listing.scss';
+import './Style.scss';
 
 
 import { Link } from "react-router-dom";
