+++ client/src/components/layout/ProfileListing.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import {Card, Button} from 'react-bootstrap'
-import '../Application.scss';
+import './Style.scss';
 import { Link } from "react-router-dom";
 import axios from 'axios'
