+++ client/src/components/layout/Style.scss
@@ -1,0 +1,296 @@
+//scss variables
+$untGreen: #006A31;
+$font-family-base:"Open Sans", Verdana, sans-serif;
+$title-font: "Gentium Book Basic", Verdana, sans-serif;
+$theme-colors: (
+    primary: $untGreen,
+);
+$enable-shadows: true;
+$enable-gradients: true;
+
+//import bootstrap scss and fonts
+@import '~bootstrap/scss/bootstrap.scss';
+@import url('https://fonts.googleapis.com/css?family=Open+Sans&display=swap');
+@import url('https://fonts.googleapis.com/css?family=Gentium+Book+Basic&display=swap');
+
+
+/* -----------------------------
+        CUSTOM CSS
+ -----------------------------*/
+//general
+    h3 {
+        padding-left: 15px;
+    }
+
+    // p {
+    //     width: 750px;
+    // }
+
+    .overflow-ellipsis {
+        text-overflow: ellipsis;
+    }
+
+    .link {
+        color: #FFF;
+    }
+
+    .link:hover {
+        color: #FFF;
+        text-decoration: none;
+    }
+    .card-title{
+        font-family: $title-font;
+    }
+
+    .title-font{
+        font-family: $title-font;
+    }
+
+    .custom-control.custom-checkbox {
+        margin-left: 15px;
+    }
+
+    .forgot-password.text-left {
+        margin-left: 15px;
+    }
+    .search {
+        padding: 15px;
+    }
+
+//landing page
+.landing-page{
+    padding: 0 40px;
+    h1 {
+        text-align: center;
+        color: $untGreen;
+        font-weight: 800;
+    }
+    /* Image */
+    img.center {
+        display: block;
+        margin: 0 auto;
+    }
+}
+
+#landingImg{
+    width: 30vw;
+}
+    
+
+/* Button */
+.btn-primary {
+    display: inline-block;
+    margin-left: 15px;
+    margin-right: 5px;
+}
+
+//     &:hover,
+//     &:active,
+//     &:visited,
+//     &:focus {
+//         background-color: lighten($color: $untGreen, $amount: 10) !important;
+//         margin-left: 15px;
+//         margin-right: 5px;
+//     }
+// }
+
+/* Tags */
+.input-tag{
+    li{
+        border-radius: 20px 0px 0px 20px;
+        margin: 5px 10px;
+        list-style: none;
+        display: flex;
+        padding: 5px 10px;
+        background-color: lighten($color: $untGreen, $amount: 20);
+
+        button{
+            border: none;
+            background-color: lighten($color: $untGreen, $amount: 20);
+            text-align: center;
+        }
+    }
+
+    ul{
+        padding-left: 0;
+    }
+}
+
+.tag-display{
+    li{
+        border-radius: 20px 0px 0px 20px;
+        margin: 5px 10px;
+        list-style: none;
+        display: flex;
+        padding: 5px 10px;
+        background-color: theme-color("light");
+    }
+
+    ul{
+        padding-left: 0;
+    }
+}
+
+/* Header/Top Nav */
+.topnav {
+    background-color: #181818;
+    overflow: hidden;
+    position: relative;
+
+    nav {
+        margin-left: 110px; 
+
+        a {
+            float: left;
+            color: #b5b5b5;
+            text-align: center;
+            padding: 14px 16px;
+            text-decoration: none;
+            font-size: 18px;
+
+            &:hover {
+                background-color: #ddd;
+                color: black;
+            }
+
+            .active {
+                color: white;
+            }
+        }
+    }
+
+    .header {
+        height: 75px;
+        position: relative;
+        background-color: $untGreen;
+    }
+
+    .banner {
+        display: block;
+        height: 120px;
+        width: auto;
+        z-index: 998;
+        position: absolute;
+        margin-left: 50px;
+
+        img.unt-banner {
+            width: auto;
+            height: 100%;
+            //padding-left: 30px;
+        }
+    }
+
+    .title-text {
+        color: white;
+        font-family: $title-font;
+        padding-left: 170px;
+        margin-top: 8px;
+        white-space: nowrap;
+
+        span {
+            padding-left: 20px;
+            font-family: $font-family-base;
+        }
+    }
+
+    .overflow {
+        white-space: nowrap;
+        overflow: hidden
+    }
+
+    .navbar {
+        height: 50px;
+        padding-left: 50px;
+    }
+
+}
+
+//Filtering projects
+.filter-projects{
+    display: flex;
+    padding-left: 20px;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 2em;
+
+    h5{
+        margin-right: 30px;
+        margin-top: auto;
+        margin-bottom: auto;
+    }
+
+    form{
+        display: flex;
+        align-items: center;
+
+        label{
+            margin-right: 20px;
+        }
+    }
+
+    button{
+        margin: 5px 0;
+    }
+}
+
+//Project card
+.project-card{
+    display: flex;
+    overflow: 1 1 auto;
+    border-radius: 20px;
+    border: none;
+    box-shadow: 8px 13px 23px -5px rgba(0,0,0,0.5);
+
+    .project-card-image{
+        display: block;
+        max-width: 30%;
+        max-height: 90%;
+        float: left;
+        margin: auto 10px;
+        img{
+            background-color: white;
+            border-radius: 20px;
+        }
+    }
+
+    .project-card-body{
+        padding: 20px;
+        width: 100%;
+    }
+
+    .view-button{
+        z-index: 3;
+        position: relative;
+        border-radius: 20px;
+        border: none;
+    }
+
+    .githubIcon{
+        border-radius: 20px;
+        margin-right: 10px;
+        width: 140px;
+        z-index: 1;
+        position: relative;
+        left: -105px;
+
+        svg{
+            float: right;
+        }
+    }
+}
+.project-card.bg-0{
+    background-color: #B9E8B7;
+}
+.project-card.bg-1{
+    background-color: #A1E6EF;
+}
+
+
+//login screen
+.form-div{
+    background-color: theme-color('light');
+    padding: 20px;
+    border-radius: 20px;
+    border: 1px solid darken(theme-color('light'),10);
+    box-shadow: 8px 13px 23px -5px rgba(0,0,0,0.5);
+}
