+++ client/src/components/redux/AddProject.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import {Container} from 'react-bootstrap'
-import CreateProject from './CreateProjectContainer';
+import CreateProject from '../../containers/CreateProject';
 
-import '../Application.scss';
+import '../layout/Style.scss';
 
 class AddProject extends Component {
   render() {
