+++ client/src/components/redux/Projects.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import ProjectList from './ProjectListContainer';
+import ProjectList from '../../containers/ProjectList';
 import ProjectSearch from '../../utils/ProjectSearch';
 import { fetchAllProjects } from '../../actions';
 import FilterProject from '../../utils/FilterProjects';
