+++ client/src/components/redux/ViewProfileProject.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import '../Application.scss';
-import ViewAProfileProject from './ViewAProfileProjectContainer';
+import '../layout/Style.scss';
+import ViewAProfileProject from '../../containers/ViewAProfileProject';
