+++ client/src/components/redux/ViewProject.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Container } from 'react-bootstrap'
 import { connect } from 'react-redux';
-import '../Application.scss';
-import ViewAProject from './ViewAProjectContainer';
+import '../layout/Style.scss';
+import ViewAProject from '../../containers/ViewAProject';
