+++ client/src/containers/CreateProject.js
@@ -1,8 +1,8 @@
 // CreateProject.js
 
 import { connect } from 'react-redux';
-import { createProject } from '../../actions';
-import AddProject from './AddProject';
+import { createProject } from '../actions';
+import AddProject from '../components/AddProject';
 
 const mapDispatchToProps = dispatch => {
   return {
